Added some suggestions
We don't describe people with disabilities as having needs. We used to, but this frames the disabed people as causign the problem. The problem isn't the disabled people, it's the inaccessible structures.

I don't think colorblindness is a contrast issue.

I think we should encourage people to provide their own captions or transcripts. This is a great time for the community to learn and take responsibility for inclusion.

I can be a contact for accessibility questions.